add svg module, implement icon usage in PodcastPlayer
